LayoutLmv1_codebase/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
attrs==24.2.0
cachetools==5.5.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
dill==0.3.8
et_xmlfile==2.0.0
filelock==3.13.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2023.10.0
fuzzywuzzy==0.18.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-vision==3.4.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.62.3
huggingface-hub==0.26.5
idna==3.4
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
lit==15.0.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
opencv-python-headless==4.8.1.78
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==10.2.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psutil==5.9.7
pyarrow==18.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seqeval==1.2.2
shapely==2.0.6
six==1.17.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tokenizers==0.15.2
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.67.1
transformers==4.36.0
triton==2.0.0
typing_extensions==4.9.0
tzdata==2024.2
urllib3==1.26.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3